// storelink-frontend/src/app/model/payment/payment.ts
export interface Payment {
  id?: number; // optional because it will be generated by backend
  cardholderName: string;
  cardNumber?: string;
  expiryMonth?: string; // format: "01" to "12"
  expiryYear?: string; // 2-digit year, e.g., "26"
  cvv?: string; // transient, not stored in backend
  otp?: string; // transient, not stored in backend
  amount: number;
  status?: 'PENDING' | 'SUCCESS' | 'FAILED'; // default PENDING
  paymentDate?: string; // ISO string set by backend
}